requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.2
django-model-utils==4.0.0
djangorestframework==3.11.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
django-environ==0.4.5
djangorestframework-simplejwt==4.4.0